backend/requirements.txt
# FastAPI and web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.8.2
python-multipart==0.0.6

# CrewAI and LLM
crewai==0.41.1
crewai-tools==0.4.26
litellm==1.44.22
openai>=1.13.3,<2.0.0


# HTTP and web tools
httpx==0.25.2
requests==2.31.0
aiohttp>=3.9.1,<4.0.0
beautifulsoup4>=4.12.3,<5.0.0
lxml==4.9.3

# Environment and utilities
python-dotenv==1.0.0
pydantic-settings==2.1.0
typing-extensions>=4.11.0,<5.0.0

# Development tools
pytest==8.3.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
